requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
supabase
pydantic
python-dotenv
